Added --keep-generated-files

+++ src/c3/create_gridwrapper.py
@@ -163,6 +163,9 @@ def main():
     parser.add_argument('--no-cache', action='store_true', help='Not using cache for docker build.')
     parser.add_argument('--skip-logging', action='store_true',
                         help='Exclude logging code from component setup code')
+    parser.add_argument('--keep-generated-files', action='store_true',
+                        help='Do not delete temporary generated files.')
+
     args = parser.parse_args()
 
     # Init logging
@@ -174,41 +177,52 @@ def main():
     handler.setLevel(args.log_level)
     root.addHandler(handler)
 
-    grid_wrapper_file_path, component_path = apply_grid_wrapper(
-        file_path=args.FILE_PATH,
-        component_process=args.component_process,
-        cos=args.cos,
-    )
+    grid_wrapper_file_path = component_path = ''
+    try:
+        grid_wrapper_file_path, component_path = apply_grid_wrapper(
+            file_path=args.FILE_PATH,
+            component_process=args.component_process,
+            cos=args.cos,
+        )
 
-    logging.info('Generate CLAIMED operator for grid wrapper')
+        logging.info('Generate CLAIMED operator for grid wrapper')
 
-    # Add component path and init file path to additional_files
-    args.ADDITIONAL_FILES.append(component_path)
+        # Add component path and init file path to additional_files
+        args.ADDITIONAL_FILES.append(component_path)
 
-    # Update dockerfile template if specified
-    if args.dockerfile_template_path != '':
-        logging.info(f'Uses custom dockerfile template from {args.dockerfile_template_path}')
-        with open(args.dockerfile_template_path, 'r') as f:
-            custom_dockerfile_template = Template(f.read())
-    else:
-        custom_dockerfile_template = None
+        # Update dockerfile template if specified
+        if args.dockerfile_template_path != '':
+            logging.info(f'Uses custom dockerfile template from {args.dockerfile_template_path}')
+            with open(args.dockerfile_template_path, 'r') as f:
+                custom_dockerfile_template = Template(f.read())
+        else:
+            custom_dockerfile_template = None
 
-    create_operator(
-        file_path=grid_wrapper_file_path,
-        repository=args.repository,
-        version=args.version,
-        custom_dockerfile_template=custom_dockerfile_template,
-        additional_files=args.ADDITIONAL_FILES,
-        log_level=args.log_level,
-        local_mode=args.local_mode,
-        no_cache=args.no_cache,
-        overwrite_files=args.overwrite,
-        rename_files=args.rename,
-        skip_logging=args.skip_logging,
-    )
-
-    logging.info('Remove local component file')
-    os.remove(component_path)
+        create_operator(
+            file_path=grid_wrapper_file_path,
+            repository=args.repository,
+            version=args.version,
+            custom_dockerfile_template=custom_dockerfile_template,
+            additional_files=args.ADDITIONAL_FILES,
+            log_level=args.log_level,
+            local_mode=args.local_mode,
+            no_cache=args.no_cache,
+            overwrite_files=args.overwrite,
+            rename_files=args.rename,
+            skip_logging=args.skip_logging,
+            keep_generated_files=args.keep_generated_files,
+        )
+    except Exception as err:
+        logging.error('Error while generating CLAIMED grid wrapper. '
+                      'Consider using `--log_level DEBUG` and `--keep-generated-files` for debugging.')
+        raise err
+    finally:
+        if not args.keep_generated_files:
+            logging.info('Remove local component file and grid wrapper code.')
+            if os.path.isfile(grid_wrapper_file_path):
+                os.remove(grid_wrapper_file_path)
+            if os.path.isfile(component_path):
+                os.remove(component_path)
 
 
 if __name__ == '__main__':
